Fragment out classes into their own files
